Upgraded to CherryPy WSGIServer 6.0.1
Necessary for Py3 support. Using the same technique as CherryPy, two
files are provided, one for Py2 and one for Py3.
